Basic Calculator
Just a basic, non GUI calculator that utilizes functions, loops and lists. It takes in two numbers with one of eight operators and spits out an answer. It uses a number of prompts and inputs to make sure a user inputs known or valid information. A full scientific calculator is in the works. Enjoy!