Re upload
Here's an entry-level function that takes in any integer, and returns its ordinal indicator. For example, if it were given the integer 1 as an input, it would return the string "st" or if it were given the integer 14, it would return "th". This can be useful for dynamically displaying or printing numbers with their proper ordinal indicators. I've included two tests after the main function. The first prints a range of numbers with their ordinal indicators (1st, 2nd, 3rd, 4th, etc...) and the second takes a user's input and adds the function with that number to the end to print the two together (27 -> "27th".) Enjoy!