2. Currency Change
A program that uses basic recursion to calculate and print the most efficient set of denominations needed for a certain dollar figure. For example, an input of $77.67 would print the statement "You're gonna need a fifty-dollar bill, a twenty-dollar bill, a five-dollar bill, 2 ones, 2 quarters, a dime, a nickel, and 2 pennies." It includes a helper function that allows a user to preference whether they'd like to use rare denominations like two-dollar bills or half-dollar coins. Enjoy!