Primes
A script that takes a given range of real, positive integers, and displays a grid of all prime or composite numbers within the range. Functions for generating both prime and composite lists have somewhat reasonable runtimes, with the composite function using the Sieve of Eratosthenes method. Unreasonably large ranges will delay the program as it uses recursive calls, and will also be difficult to distinguish on the display. Enjoy!